Reintroduce memoization.
Memoization is not a simple matter: A function can appear inside
two .su files - we'd store and reuse the same value for both places
(which is wrong). But we do need to do memoization, otherwise
deep call stacks - like the ones used in RTEMS libraries - will
make the runtime of this utility explode...

So indeed do perform memoization in findStackUsage; but only if
ALL symbols in the call chain leading to us appear uniquely inside
the ELF (see static functions counter-example in the code).